support custom file paths

+++ Library/T4FileGenerator/Services/FileManager.cs
@@ -25,18 +25,17 @@ namespace T4FileGenerator.Services
 			_configuration = configuration;
 		}
 
-		public List<OutputFile> LoadFiles(string subPath = null, IEnumerable<string> excludedDirs = null)
+		public List<OutputFile> LoadFiles(string path = null, IEnumerable<string> excludedDirs = null)
 		{
 			var excludeDirs = excludedDirs != null ? excludedDirs.Concat(_excludedDirs).ToList() : _excludedDirs;
 			var directoryWalker = new DirectoryWalker(excludeDirs);
-			var root = subPath != null ? Path.Combine(_configuration.TargetPath, subPath) : _configuration.TargetPath;
-			var fileNames = directoryWalker.WalkDownDirectory(root);
+			var fileNames = directoryWalker.WalkDownDirectory(path);
 			return fileNames
 				.Select(fileName => new OutputFile
 				{
 					Contents = File.ReadAllText(fileName),
 					FileName = fileName,
-					RelativeFileName = Path.GetRelativePath(root, fileName)
+					RelativeFileName = Path.GetRelativePath(path, fileName)
 				})
 				.ToList();
 		}
@@ -48,8 +47,8 @@ namespace T4FileGenerator.Services
 				return false;
 			}
 
-			return StringHelpers.NormalizePath(Path.Combine(_configuration.TargetPath, path1)) ==
-			       StringHelpers.NormalizePath(Path.Combine(_configuration.TargetPath, path2));
+			return StringHelpers.NormalizePath(path1) ==
+			       StringHelpers.NormalizePath(path2);
 		}
 
 		public List<OutputFile> ComputeOutputFiles()
@@ -60,8 +59,7 @@ namespace T4FileGenerator.Services
 			{
 				var humanFile =
 					HumanFiles.FirstOrDefault(x => ComparePaths(x.RelativeFileName, generatedFile.RelativeFileName));
-				var originalFile = OriginalFiles.FirstOrDefault(x =>
-					ComparePaths(x.RelativeFileName, generatedFile.RelativeFileName));
+				var originalFile = OriginalFiles.FirstOrDefault(x => ComparePaths(x.RelativeFileName, generatedFile.RelativeFileName));
 				// if overwrite human, then overwrite human
 				if (_configuration.OverwriteHuman)
 				{
@@ -92,8 +90,8 @@ namespace T4FileGenerator.Services
 			foreach (var outputFile in outputFiles)
 			{
 				var outputPath = subPath != null
-					? Path.Combine(_configuration.TargetPath, subPath, outputFile.RelativeFileName)
-					: Path.Combine(_configuration.TargetPath, outputFile.RelativeFileName);
+					? Path.Combine(subPath, outputFile.RelativeFileName)
+					: Path.Combine(outputFile.RelativeFileName);
 				var fileInfo = new FileInfo(outputPath);
 
 				if (fileInfo.Directory != null && !fileInfo.Exists)
@@ -105,16 +103,16 @@ namespace T4FileGenerator.Services
 			}
 		}
 
-		public void CleanOldGeneratedFiles()
+		public void CleanOldGeneratedFiles(string path, List<string>? cleaningExclude = null)
 		{
 			foreach (var originalFile in OriginalFiles)
 			{
 				// if it existed as an original file but not a newly generated one
 				var generatedFile = GeneratedFiles.FirstOrDefault(x => ComparePaths(x.RelativeFileName, originalFile.RelativeFileName));
-				if (generatedFile == null)
+				if (generatedFile == null && cleaningExclude is not null && !cleaningExclude.Contains(originalFile.FileName))
 				{
 					File.Delete(originalFile.FileName);
-					File.Delete(Path.Combine(_configuration.TargetPath, originalFile.RelativeFileName));
+					File.Delete(Path.Combine(path, originalFile.RelativeFileName));
 				}
 			}
 		}
